# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Media Server API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Library {
  id: number;
  name: string;
  path: string;
  createdAt: string;
  lastScannedAt?: string | null;
}

export interface AddLibraryRequest {
  name: string;
  path: string;
}

export interface ScanStatus {
  scanning: boolean;
  currentLibraryId?: number | null;
  tracksScanned: number;
  tracksAdded: number;
  tracksUpdated: number;
  tracksRemoved: number;
  startedAt?: string | null;
}

export interface Artist {
  id: number;
  name: string;
  albumCount: number;
  trackCount: number;
}

export interface Album {
  id: number;
  title: string;
  artistId?: number | null;
  artistName?: string | null;
  year?: number | null;
  trackCount: number;
  hasArt: boolean;
  genre?: string | null;
}

export interface ArtistDetail {
  id: number;
  name: string;
  albumCount: number;
  trackCount: number;
  albums: Album[];
}

export interface ArtistListResponse {
  artists: Artist[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Track {
  id: number;
  title: string;
  artistId?: number | null;
  artistName?: string | null;
  albumId?: number | null;
  albumTitle?: string | null;
  trackNumber?: number | null;
  discNumber?: number | null;
  durationSeconds?: number | null;
  genre?: string | null;
  year?: number | null;
  filePath: string;
  mimeType: string;
  hasArt: boolean;
}

export interface AlbumDetail {
  id: number;
  title: string;
  artistId?: number | null;
  artistName?: string | null;
  year?: number | null;
  trackCount: number;
  hasArt: boolean;
  genre?: string | null;
  tracks: Track[];
}

export interface AlbumListResponse {
  albums: Album[];
  total: number;
  page: number;
  pageSize: number;
}

export interface TrackListResponse {
  tracks: Track[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Genre {
  name: string;
  trackCount: number;
}

export type FolderEntryType =
  (typeof FolderEntryType)[keyof typeof FolderEntryType];

export const FolderEntryType = {
  directory: "directory",
  file: "file",
} as const;

export interface FolderEntry {
  name: string;
  path: string;
  type: FolderEntryType;
  trackId?: number | null;
  mimeType?: string | null;
}

export interface FolderContents {
  path: string;
  entries: FolderEntry[];
}

export type ListLibraries200 = {
  libraries: Library[];
};

export type ListArtistsParams = {
  page?: number;
  pageSize?: number;
  search?: string;
};

export type GetArtistAlbums200 = {
  albums: Album[];
};

export type ListAlbumsParams = {
  page?: number;
  pageSize?: number;
  search?: string;
  artistId?: number;
};

export type GetAlbumTracks200 = {
  tracks: Track[];
};

export type ListTracksParams = {
  page?: number;
  pageSize?: number;
  search?: string;
  albumId?: number;
  artistId?: number;
};

export type ListGenres200 = {
  genres: Genre[];
};

export type GetGenreTracksParams = {
  page?: number;
  pageSize?: number;
};

export type BrowseFolderParams = {
  path?: string;
};