# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Media Server API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: number;
  title: string;
  artistId?: number | null;
  artistName?: string | null;
  year?: number | null;
  trackCount: number;
  hasArt: boolean;
  genre?: string | null;
}