# lib/api-zod/src/generated/types/artistDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Media Server API
 * OpenAPI spec version: 0.1.0
 */
import type { Album } from "./album";

export interface ArtistDetail {
  id: number;
  name: string;
  albumCount: number;
  trackCount: number;
  albums: Album[];
}