# lib/api-zod/src/generated/types/library.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Media Server API
 * OpenAPI spec version: 0.1.0
 */

export interface Library {
  id: number;
  name: string;
  path: string;
  createdAt: Date;
  lastScannedAt?: Date | null;
}