# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Media Server API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  id: number;
  title: string;
  artistId?: number | null;
  artistName?: string | null;
  albumId?: number | null;
  albumTitle?: string | null;
  trackNumber?: number | null;
  discNumber?: number | null;
  durationSeconds?: number | null;
  genre?: string | null;
  year?: number | null;
  filePath: string;
  mimeType: string;
  hasArt: boolean;
}